+++ ravenpy/models/__init__.py
@@ -3,7 +3,7 @@ import os
 from .base import Ostrich, Raven
 from .emulators import *
 from .multimodel import RavenMultiModel
-from .rv import RV, RVI
+from .rv import RV, RVI, HRU, HRUState, LU
 
 _dir = os.path.abspath(os.path.dirname(__file__))
 
@@ -12,6 +12,7 @@ raven_templates = {
     "raven-mohyse": os.path.join(_dir, "raven-mohyse"),
     "raven-hmets": os.path.join(_dir, "raven-hmets"),
     "raven-hbv-ec": os.path.join(_dir, "raven-hbv-ec"),
+    "raven-blended": os.path.join(_dir, "raven-blended"),
 }
 
 ostrich_templates = {
@@ -19,4 +20,13 @@ ostrich_templates = {
     "ostrich-mohyse": os.path.join(_dir, "ostrich-mohyse"),
     "ostrich-hmets": os.path.join(_dir, "ostrich-hmets"),
     "ostrich-hbv-ec": os.path.join(_dir, "ostrich-hbv-ec"),
+    "ostrich-blended": os.path.join(_dir, "ostrich-blended"),
 }
+
+grid_weight_importer_params = dict(
+    DIM_NAMES=("lon_dim", "lat_dim"),
+    VAR_NAMES=("lon", "lat"),
+    ROUTING_ID_FIELD="HRU_ID",
+    NETCDF_INPUT_FIELD="NetCDF_col",
+    AREA_ERROR_THRESHOLD=0.05,
+)