+++ tests/test_rv.py
@@ -1,25 +1,18 @@
 import datetime as dt
 import re
 from collections import namedtuple
-from io import StringIO
 from pathlib import Path
 
-import geopandas
 import pytest
 
 import ravenpy
-from ravenpy.models.commands import GriddedForcingCommand
-from ravenpy.models.importers import (
-    RoutingProductGridWeightImporter,
-    RoutingProductShapefileImporter,
-)
-from ravenpy.models.rv import (
+from ravenpy.models.commands import GriddedForcingCommand, BaseValueCommand, RainCorrection
+from ravenpy.models.rv import (  # RVT,
     RV,
     RVC,
     RVH,
     RVI,
     RVP,
-    RVT,
     MonthlyAverage,
     Ost,
     RavenNcData,
@@ -229,12 +222,14 @@ class TestRVC:
 
 
 class TestRVH:
+    importers = pytest.importorskip("ravenpy.models.importers")
+
     @classmethod
     def setup_class(self):
         shp = get_local_testdata("raven-routing-sample/finalcat_hru_info.zip")
-        importer = RoutingProductShapefileImporter(shp)
+        importer = self.importers.RoutingProductShapefileImporter(shp)
         config = importer.extract()
-        config.pop('channel_profiles')
+        config.pop("channel_profiles")
         self.rvh = RVH(**config)
 
     def test_import_process(self):
@@ -269,12 +264,14 @@ class TestRVH:
 
 
 class TestRVP:
+    importers = pytest.importorskip("ravenpy.models.importers")
+
     @classmethod
     def setup_class(self):
         shp = get_local_testdata("raven-routing-sample/finalcat_hru_info.zip")
-        importer = RoutingProductShapefileImporter(shp)
+        importer = self.importers.RoutingProductShapefileImporter(shp)
         config = importer.extract()
-        self.rvp = RVP(channel_profiles=config['channel_profiles'])
+        self.rvp = RVP(channel_profiles=config["channel_profiles"])
 
     def test_import_process(self):
         assert len(self.rvp.channel_profiles) == 46
@@ -287,11 +284,15 @@ class TestRVP:
 
 
 class TestRVT:
+    importers = pytest.importorskip("ravenpy.models.importers")
+
     @classmethod
     def setup_class(self):
         input_file = get_local_testdata("raven-routing-sample/VIC_streaminputs.nc")
         routing_file = get_local_testdata("raven-routing-sample/finalcat_hru_info.zip")
-        importer = RoutingProductGridWeightImporter(input_file, routing_file)
+        importer = self.importers.RoutingProductGridWeightImporter(
+            input_file, routing_file
+        )
         gws = importer.extract()
         self.gfc = GriddedForcingCommand(grid_weights=gws)
 
@@ -300,7 +301,8 @@ class TestRVT:
 
         assert ":NumberHRUs 51" in res
         assert ":NumberGridCells 100" in res
-        assert len(res.split("\n")) == 227
+        # FIXME: This test is not superb.
+        assert len(res.split("\n")) == 225
 
 
 def test_isinstance_namedtuple():
@@ -308,3 +310,9 @@ def test_isinstance_namedtuple():
     x = X(1, 2, 3)
     assert isinstance_namedtuple(x)
     assert not isinstance_namedtuple([1, 2, 3])
+
+
+class TestBaseValueCommand:
+    def test_raincorrection(self):
+        rc = RainCorrection(3)
+        assert f"{rc}" == ":RainCorrection 3"